A Research Project Question
This problem asks the user to the amount of Facebook profiles, the number of pictures on each profile, followed by the rating of attractiveness on each picture to determine the right fit for the subject. But unfortunately, Facebook's user count has been declining so this problem would not fit in a real-world scenario.